Rename DownloadEntities class
